backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.34.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv==1.0.0
markdown==3.7
python-multipart==0.0.9
sse-starlette==2.1.0
aiofiles==24.1.0
